$.BibEntry {RefManageR}	R Documentation

Extract fields from a BibEntry object

Description

used to extract a single field from each entry in a BibEntry object

Usage

## S3 method for class 'BibEntry'
x$name

Arguments

x
an object of class BibEntry

name
the field to extract

Value

a named list of values for the field specified by name for each entry; NULL if the field is not present for a particular entry. The names attribute of the returned list contains the entry keys (potentially back-quoted).

Note

name may be “bibtype” to extract entry types or “key” to extract keys.

See Also

Other operators: $<-.BibEntry(), +.BibEntry(), [.BibEntry(), [<-.BibEntry(), [[.BibEntry(), [[<-.BibEntry(), c.BibEntry()

Examples

if (requireNamespace("bibtex")) {
    file.name <- system.file("Bib", "biblatexExamples.bib", package="RefManageR")
    bib <- suppressMessages(ReadBib(file.name))
    bib[[50:55]]$author
    bib[[seq_len(5)]]$bibtype
 }

[Package RefManageR version 1.3.0 Index]