names<-.BibEntry {RefManageR}	R Documentation

Names (keys) of a BibEntry object

Description

Functions to get and set the keys of an object of class BibEntry

Usage

## S3 replacement method for class 'BibEntry'
names(x) <- value

## S3 method for class 'BibEntry'
names(x)

Arguments

x
an object of class BibEntry

value
character vector of new key values to replace into x

Value

names<- the updated BibEntry object.

names - character vector of the keys of the BibEntry object.

Author(s)

McLean, M. W. mathew.w.mclean@gmail.com

Examples

if (requireNamespace("bibtex")) {
    bib <- ReadBib(system.file("Bib", "test.bib", package = "RefManageR"))
    names(bib)
    names(bib)[1] <- 'newkey'
}

[Package RefManageR version 1.3.0 Index]