[[<-.BibEntry {RefManageR}	R Documentation

Assign a BibEntry entry to another BibEntry object

Description

Replace one entry in a BibEntry object with another

Usage

## S3 replacement method for class 'BibEntry'
x[[i]] <- value

Arguments

x
- a BibEntry object

i
- a numeric index or a string entry key

value
- a single entry BibEntry object or an object that can be coerced to BibEntry using as.BibEntry

Details

This function will replace the specified entry in x with the entry given by value. To replace multiple entries see [<-.BibEntry.

Value

an object of class BibEntry

See Also

Other operators: $.BibEntry(), $<-.BibEntry(), +.BibEntry(), [.BibEntry(), [<-.BibEntry(), [[.BibEntry(), c.BibEntry()

[Package RefManageR version 1.3.0 Index]