Replace values for a particular field in a BibEntry object

Used to replace the values stored for a specified field in a BibEntry object.

# S3 method for BibEntry
$(x, name) <- value

Arguments

x
a BibEntry object

name
string; the field to assign the new values to.

value
character vector; the replacement field values to be assigned.

Value

an object of class BibEntry with the updated fields.

Note

The method expects date and name list fields to be in the format expected by Biblatex. The field specified by name does not have to be one currently in x.

See also

Examples

bib <- BibEntry(bibtype = "misc", key = "mclean", author = "Mathew W. McLean", 
  title = "My Work", year = "2012")
bib$year <- 2014
bib$author <- "McLean, M. W. and Carroll, R. J." 
bib$url <- "https://example.com"
bib

#> [1] M. W. McLean and R. J. Carroll. _My Work_. 2014. <URL:
#> https://example.com>.

bib <- c(bib, as.BibEntry(citation()))
bib[1]$author[2] <- person(c("Raymond", "J."), "Carroll")

#> Error in (function (x, i, j, ..., drop = FALSE) {    if (!length(x) || (missing(i) && missing(...)))         return(x)    ind <- NULL    if (missing(i)) {        dots <- list(...)    }    else if (is.numeric(i)) {        ind <- i    }    else if (is.logical(i)) {        ind <- which(i)    }    else if (is.character(i) && missing(j) && missing(...)) {        if (is.null(names(i))) {            ind <- match(i, names(x))            ind <- ind[!is.na(ind)]        }        else {            dots <- as.list(i)        }    }    else if (is.list(i) && missing(j) && missing(...)) {        dots <- i    }    else if (is.list(i) || is.character(i)) {        kall <- match.call(expand.dots = FALSE)        if (!missing(j)) {            kall$j <- NULL        }        if (!missing(...)) {            kall$... <- NULL        }        if (is.list(i[[1L]])) {            kall$i <- i[[1L]]            kall$j <- i[[-1L]]        }        ret.ind <- .BibOptions$return.ind        .BibOptions$return.ind <- TRUE        tryCatch({            kall$x <- force(x)            ind <- suppressMessages(eval(kall))            if (!missing(j)) {                if (is.list(j[[1L]])) {                  kall$i <- j[[1L]]                  kall$j <- j[[-1L]]                }                else {                  kall$i <- j                  if (!missing(...)) {                    kall$j <- list(...)                  }                }                ind <- unique(c(ind, suppressMessages(eval(kall))))            }            else if (!missing(...)) {                kall$i <- list(...)                ind <- unique(c(ind, suppressMessages(eval(kall))))            }        }, error = function(e) {            .BibOptions$return.ind <- ret.ind            stop(e)        })        .BibOptions$return.ind <- ret.ind    }    else {        stop("Invalid index.")    }    if (exists("dots", inherits = FALSE)) {        add <- function(x) suppressMessages(Reduce("|", x))        y <- .BibEntry_expand_crossrefs(x)        fields <- names(dots)        ind <- seq_along(x)        for (i in seq_along(dots)) {            ind <- ind[add(lapply(dots[[i]], function(trm, bib,                 fld) {                len <- nchar(trm)                if (len > 2L && substr(trm, 1L, 1L) == "!") {                  trm <- substr(trm, 2L, len)                  !FindBibEntry(bib, trm, fld)                }                else {                  FindBibEntry(bib, trm, fld)                }            }, bib = y[[ind]], fld = fields[i]))]            if (!length(ind))                 break        }    }    if (.BibOptions$return.ind)         return(ind)    if (!length(ind)) {        message("No results.")        return(invisible(list()))    }    y <- .BibEntry_expand_crossrefs(unclass(x[[ind]]), unclass(x[[-ind]]))    if (!drop)         attributes(y) <- attributes(x)[bibentry_list_attribute_names]    class(y) <- c("BibEntry", "bibentry")    return(y)})(x = `*tmp*`, i = 1): object '*tmp*' not found

bib$author

#> $mclean
#> [1] "M. W. McLean"  "R. J. Carroll"
#> 
#> $`2021language`
#> [1] "R Core Team"
#>