print.BibEntry {RefManageR}	R Documentation

Print BibLaTeX bibliography Entries

Description

Prints bibliographic information stored in BibEntry objects in BibLaTeX style

Usage

## S3 method for class 'BibEntry'
print(x, .opts = list(), ...)

Arguments

x
a BibEntry object

.opts
a list of formatting options from BibOptions. Possible options are

style - character string naming the printing style. Possible values are plain text (style “text”), BibTeX (“Bibtex”), BibLaTeX (“Biblatex”), a mixture of plain text and BibTeX as traditionally used for citations (“citation”), HTML (“html”), LaTeX (“latex”), “markdown”, “yaml”, R code (“R”), and a simple copy of the textVersion elements (style “textVersion”, see BibEntry)

bib.style - character string specifying BibLaTeX style to use for formatting references. Possible values are “numeric” (default), “authoryear”, “authortitle”, “alphabetic”, “draft”. See section 3.3.2 of the BibLaTeX manual.

sorting - how should the entries in x be sorted? See sort.BibEntry.

max.names - maximum number of names to display for name list fields before truncation with “et al.”.

first.inits - logical; if true only initials of given names are printed, otherwise full names are used.

dashed - logical; for .bibstyle = "authoryear" or .bibstyle = "authoryear" only, if TRUE duplicate author and editor lists are replaced with “—” when printed.

no.print.fields character vector; fields that should not be printed, e.g., doi, url, isbn, etc.

...
extra parameters to pass to the renderer.

Note

setting max.names to value is equivalent to setting maxnames=value and minnames=value in BibLaTeX.

Custom BibLaTeX styles may be defined using the function bibstyle. To fully support BibLaTeX, the created environment must have functions for formatting each of the entry types described in BibEntry.

References

Lehman, Philipp and Kime, Philip and Boruvka, Audrey and Wright, J. (2013). The biblatex Package. https://mirror.pregi.net/tex-archive/macros/latex/contrib/biblatex/doc/biblatex.pdf.

See Also

BibEntry, ReadBib, sort.BibEntry

Examples

if (requireNamespace("bibtex")) {
    file.name <- system.file("Bib", "biblatexExamples.bib", package="RefManageR")
    bib <- suppressMessages(ReadBib(file.name))
    print(bib[author="aristotle"], .opts = list(bib.style = "numeric"))
    print(bib[55:57], .opts = list(bib.style = "authortitle", first.inits = FALSE))
    print(bib[80:88], .opts = list(bib.style = "alphabetic", max.names = 1,
          no.print.fields = "issn"))
    print(bib[32:36], .opts = list(bib.style = "draft"))
    oldopts <- BibOptions(bib.style = "authoryear", dashed = TRUE, sorting = "ydnt")
    bib[editor = "westfahl"]
    BibOptions(oldopts)
}

[Package RefManageR version 1.3.0 Index]